clients/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "django-openapi-gateway"
version = "0.1.0"
description = "Django middleware for services running behind openapi-gateway"
requires-python = ">=3.11"
dependencies = [
    "django>=4.2",
]

[project.optional-dependencies]
drf = ["djangorestframework>=3.14"]

[tool.setuptools.packages.find]
where = ["."]
include = ["django_gateway*"]
